+++ packages/cloud_firestore/cloud_firestore/android/src/main/java/io/flutter/plugins/firebase/firestore/GeneratedAndroidFirebaseFirestore.java
@@ -166,6 +166,22 @@ public class GeneratedAndroidFirebaseFirestore {
     }
   }
 
+  /**
+   * [PersistenceCacheIndexManagerRequest] represents the request types for the persistence cache
+   * index manager.
+   */
+  public enum PersistenceCacheIndexManagerRequest {
+    ENABLE_INDEX_AUTO_CREATION(0),
+    DISABLE_INDEX_AUTO_CREATION(1),
+    DELETE_ALL_INDEXES(2);
+
+    final int index;
+
+    private PersistenceCacheIndexManagerRequest(final int index) {
+      this.index = index;
+    }
+  }
+
   public enum PigeonTransactionResult {
     SUCCESS(0),
     FAILURE(1);
@@ -1815,6 +1831,11 @@ public class GeneratedAndroidFirebaseFirestore {
         @NonNull ListenSource source,
         @NonNull Result<String> result);
 
+    void persistenceCacheIndexManagerRequest(
+        @NonNull FirestorePigeonFirebaseApp app,
+        @NonNull PersistenceCacheIndexManagerRequest request,
+        @NonNull Result<Void> result);
+
     /** The codec used by FirebaseFirestoreHostApi. */
     static @NonNull MessageCodec<Object> getCodec() {
       return FirebaseFirestoreHostApiCodec.INSTANCE;
@@ -2565,6 +2586,39 @@ public class GeneratedAndroidFirebaseFirestore {
 
                 api.documentReferenceSnapshot(
                     appArg, parametersArg, includeMetadataChangesArg, sourceArg, resultCallback);
+              });
+        } else {
+          channel.setMessageHandler(null);
+        }
+      }
+      {
+        BasicMessageChannel<Object> channel =
+            new BasicMessageChannel<>(
+                binaryMessenger,
+                "dev.flutter.pigeon.cloud_firestore_platform_interface.FirebaseFirestoreHostApi.persistenceCacheIndexManagerRequest",
+                getCodec());
+        if (api != null) {
+          channel.setMessageHandler(
+              (message, reply) -> {
+                ArrayList<Object> wrapped = new ArrayList<Object>();
+                ArrayList<Object> args = (ArrayList<Object>) message;
+                FirestorePigeonFirebaseApp appArg = (FirestorePigeonFirebaseApp) args.get(0);
+                PersistenceCacheIndexManagerRequest requestArg =
+                    PersistenceCacheIndexManagerRequest.values()[(int) args.get(1)];
+                Result<Void> resultCallback =
+                    new Result<Void>() {
+                      public void success(Void result) {
+                        wrapped.add(0, null);
+                        reply.reply(wrapped);
+                      }
+
+                      public void error(Throwable error) {
+                        ArrayList<Object> wrappedError = wrapError(error);
+                        reply.reply(wrappedError);
+                      }
+                    };
+
+                api.persistenceCacheIndexManagerRequest(appArg, requestArg, resultCallback);
               });
         } else {
           channel.setMessageHandler(null);