+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/method_channel_document_reference.dart
@@ -110,7 +110,7 @@ class MethodChannelDocumentReference extends DocumentReferencePlatform {
     bool includeMetadataChanges = false,
     ServerTimestampBehavior serverTimestampBehavior =
         ServerTimestampBehavior.none,
-    ListenSource source = ListenSource.defaultSource,
+    required ListenSource listenSource,
   }) {
     // It's fine to let the StreamController be garbage collected once all the
     // subscribers have cancelled; this analyzer warning is safe to ignore.
@@ -128,7 +128,7 @@ class MethodChannelDocumentReference extends DocumentReferencePlatform {
             serverTimestampBehavior: serverTimestampBehavior,
           ),
           includeMetadataChanges,
-          source,
+          listenSource,
         );
         snapshotStreamSubscription =
             MethodChannelFirebaseFirestore.documentSnapshotChannel(observerId)