+++ packages/firebase_core/firebase_core/example/android/app/src/main/kotlin/io/flutter/plugins/firebasecoreexample/MainActivity.kt
@@ -1,0 +1,5 @@
+package io.flutter.plugins.firebasecoreexample
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()